Refactor import paths and enhance EyeClassifier and EyeDeserializer classes for improved organization
This commit updates the import paths in the EyeClassifier and EyeDeserializer classes, transitioning to a new structure that enhances clarity and maintainability. Additionally, it introduces detailed JSDoc comments for methods in the EyeClassifier class, improving documentation and understanding of the codebase. New types for eye-tracking data files and settings have also been added, further streamlining the data processing functionality.
