+++ src/lib/gaze-data/back-process/class/EyeClassifier/EyeClassifier.ts
@@ -1,7 +1,18 @@
-import type { EyeFileType } from '$lib/type/EyeFile/EyeFileType.js'
-import type { EyeSettingsType } from '$lib/type/Settings/EyeSettings/EyeSettingsType.js'
+import type { EyeFileType } from '$lib/gaze-data/back-process/types/EyeFileType.js'
+import type { EyeSettingsType } from '$lib/gaze-data/back-process/types/EyeSettingsType.js'
 
+/**
+ * Class responsible for identifying and classifying eye-tracking data file formats.
+ * This classifier analyzes file content to determine the appropriate parsing settings
+ * for different eye-tracking data formats (Tobii, GazePoint, BeGaze, etc.).
+ */
 export class EyeClassifier {
+  /**
+   * Analyzes a slice of file content to determine the appropriate parsing settings.
+   * @param slice - A portion of the file content to analyze
+   * @returns EyeSettingsType object containing the determined parsing settings
+   * @throws Error if the file type cannot be determined
+   */
   classify(slice: string): EyeSettingsType {
     const type = this.getTypeFromSlice(slice)
     if (type === 'unknown') throw new Error('Unknown file type')
@@ -17,6 +28,11 @@ export class EyeClassifier {
     }
   }
 
+  /**
+   * Determines the eye-tracking data file type from a content slice.
+   * @param slice - A portion of the file content to analyze
+   * @returns The identified EyeFileType or 'unknown' if type cannot be determined
+   */
   getTypeFromSlice(slice: string): EyeFileType {
     if (this.isTobii(slice)) {
       if (slice.includes('Event')) return 'tobii-with-event'
@@ -31,6 +47,13 @@ export class EyeClassifier {
     return 'unknown'
   }
 
+  /**
+   * Determines the appropriate column delimiter based on file type and content.
+   * @param type - The identified eye-tracking file type
+   * @param slice - A portion of the file content to analyze
+   * @returns The appropriate column delimiter character
+   * @throws Error if the file type is unknown
+   */
   getColumnDelimiter(type: EyeFileType, slice: string): string {
     switch (type) {
       case 'tobii':
@@ -50,14 +73,29 @@ export class EyeClassifier {
     }
   }
 
+  /**
+   * Checks if the content matches Tobii file format.
+   * @param slice - A portion of the file content to analyze
+   * @returns true if the content matches Tobii format
+   */
   isTobii(slice: string): boolean {
     return slice.includes('Recording timestamp')
   }
 
+  /**
+   * Checks if the content matches GazePoint file format.
+   * @param slice - A portion of the file content to analyze
+   * @returns true if the content matches GazePoint format
+   */
   isGazePoint(slice: string): boolean {
     return slice.includes('FPOGS') && slice.includes('FPOGD')
   }
 
+  /**
+   * Checks if the content matches BeGaze file format.
+   * @param slice - A portion of the file content to analyze
+   * @returns true if the content matches BeGaze format
+   */
   isBeGaze(slice: string): boolean {
     return (
       slice.includes('Event Start Trial Time [ms]') &&
@@ -65,14 +103,29 @@ export class EyeClassifier {
     )
   }
 
+  /**
+   * Checks if the content matches OGAMA file format.
+   * @param slice - A portion of the file content to analyze
+   * @returns true if the content matches OGAMA format
+   */
   isOgama(slice: string): boolean {
     return slice.includes('# Contents: Similarity Measurements of scanpaths.')
   }
 
+  /**
+   * Checks if the content matches Varjo file format.
+   * @param slice - A portion of the file content to analyze
+   * @returns true if the content matches Varjo format
+   */
   isVarjo(slice: string): boolean {
     return slice.includes('Time') && slice.includes('Actor Label')
   }
 
+  /**
+   * Checks if the content matches standard CSV file format.
+   * @param slice - A portion of the file content to analyze
+   * @returns true if the content matches standard CSV format
+   */
   isCsv(slice: string): boolean {
     const header = slice.split('\r\n')[0]
     const headerColumns = header.split(',')
@@ -84,6 +137,11 @@ export class EyeClassifier {
     )
   }
 
+  /**
+   * Checks if the content matches segmented CSV file format.
+   * @param slice - A portion of the file content to analyze
+   * @returns true if the content matches segmented CSV format
+   */
   isCsvSegmented(slice: string): boolean {
     return (
       slice.includes('From') &&
