+++ src/lib/gaze-data/back-process/class/EyeDeserializer/BeGazeEyeDeserializer.ts
@@ -1,4 +1,4 @@
-import type { SingleDeserializerOutput } from '$lib/type/DeserializerOutput/SingleDeserializerOutput/SingleDeserializerOutput'
+import type { SingleDeserializerOutput } from '$lib/gaze-data/back-process/types/SingleDeserializerOutput'
 import { AbstractEyeDeserializer } from './AbstractEyeDeserializer'
 
 export class BeGazeEyeDeserializer extends AbstractEyeDeserializer {
