+++ src/lib/gaze-data/back-process/class/EyeDeserializer/CsvEyeDeserializer.ts
@@ -1,4 +1,4 @@
-import type { SingleDeserializerOutput } from '$lib/type/DeserializerOutput/SingleDeserializerOutput/SingleDeserializerOutput.js'
+import type { SingleDeserializerOutput } from '$lib/gaze-data/back-process/types/SingleDeserializerOutput.js'
 import { AbstractEyeDeserializer } from './AbstractEyeDeserializer'
 
 // BEWARE! If only one timestamp for whole segment, start and end are the same!
