+++ src/lib/gaze-data/back-process/class/EyePipeline/EyePipeline.ts
@@ -8,9 +8,9 @@ import { OgamaEyeDeserializer } from '../EyeDeserializer/OgamaEyeDeserializer'
 import { VarjoEyeDeserializer } from '../EyeDeserializer/VarjoEyeDeserializer'
 import { CsvEyeDeserializer } from '../EyeDeserializer/CsvEyeDeserializer'
 import { TobiiEyeDeserializer } from '../EyeDeserializer/TobiiEyeDeserializer'
-import type { EyeSettingsType } from '$lib/type/Settings/EyeSettings/EyeSettingsType.js'
+import type { EyeSettingsType } from '$lib/gaze-data/back-process/types/EyeSettingsType.js'
 import { EyeSplitter } from '../EyeSplitter/EyeSplitter'
-import type { DataType } from '$lib/type/Data'
+import type { DataType } from '$lib/gaze-data/shared/types'
 import { EyeRefiner } from '../EyeRefiner/EyeRefiner'
 import { CsvSegmentedEyeDeserializer } from '../EyeDeserializer/CsvSegmentedEyeDeserializer'
