+++ src/lib/gaze-data/shared/types/index.ts
@@ -1,4 +1,39 @@
 /**
+ * Used for stimuli and participants basic information.
+ */
+export interface BaseInterpretedDataType {
+  id: number
+  originalName: string
+  displayedName: string
+}
+
+/**
+ * Used for AOI and category basic information.
+ */
+export interface ExtendedInterpretedDataType extends BaseInterpretedDataType {
+  color: string
+}
+
+export interface SegmentInterpretedDataType {
+  id: number
+  start: number
+  end: number
+  category: ExtendedInterpretedDataType
+  aoi: ExtendedInterpretedDataType[]
+}
+
+export interface AttributeDataType {
+  data: string[][]
+  orderVector: number[] | []
+}
+
+export interface ParticipantsGroup {
+  id: number
+  name: string
+  participantsIds: number[]
+}
+
+/**
  * An object that represents the visibility blocks for AOIs.
  *
  * The object has string keys in the form of AAAxBBBxCCC, where AAA is the stimulusId, BBB is the aoiId,
@@ -22,4 +57,24 @@ export interface AoiDataType {
   data: string[][][]
   orderVector: number[][] | []
   dynamicVisibility: VisibilityAoiDataType
+}
+
+/**
+ * Data for workplace are stored in this unique format.
+ *
+ * It has complicated structure, but it is more performant
+ * than using, e.g., IndexedDB (up to 10x faster in Safari with 50k+ segments).
+ *
+ * This allows quick write and read operations as everything is stored in memory.
+ *
+ * @property {boolean} isOrdinalOnly - Whether the const contains only ordinal const.
+ */
+export interface DataType {
+  isOrdinalOnly: boolean
+  aois: AoiDataType
+  categories: AttributeDataType
+  participants: AttributeDataType
+  participantsGroups: ParticipantsGroup[]
+  stimuli: AttributeDataType
+  segments: number[][][][]
 }
