+++ src/lib/modals/export/class/ScanGraphDownloader.ts
@@ -5,7 +5,7 @@ import {
   getParticipant,
   getSegment,
 } from '$lib/gaze-data/front-process/stores/dataStore'
-import type { ExtendedInterpretedDataType } from '$lib/type/Data'
+import type { ExtendedInterpretedDataType } from '$lib/gaze-data/shared/types'
 
 export class ScanGraphDownloader extends AbstractDownloader {
   download(stimulusId: number, fileName: string): void {
