+++ src/lib/plots/transition-matrix/types/index.ts
@@ -1,4 +1,4 @@
-import type { ExtendedInterpretedDataType } from '$lib/type/Data'
+import type { ExtendedInterpretedDataType } from '$lib/gaze-data/shared/types'
 
 /**
  * Interface representing an Transition Matrix with labels
