--- src/lib/type/Settings/EyeSettings/EyeSettingsType.ts
@@ -1,7 +1,0 @@
-export interface EyeSettingsType {
-  rowDelimiter: string
-  columnDelimiter: string
-  type: string
-  userInputSetting: string
-  headerRowId: number
-}
