+++ tests/CsvEyeDeserializer.test.ts
@@ -5,7 +5,7 @@
  * @see src/lib/class/Eye/EyeDeserializer/CsvEyeDeserializer.ts
  */
 
-import type { SingleDeserializerOutput } from '$lib/type/DeserializerOutput/SingleDeserializerOutput/SingleDeserializerOutput'
+import type { SingleDeserializerOutput } from '$lib/gaze-data/back-process/types/SingleDeserializerOutput'
 import { CsvEyeDeserializer } from '../src/lib/class/Eye/EyeDeserializer/CsvEyeDeserializer'
 import { test, expect, describe } from 'vitest'
